requirements.txt
ultralytics>=8.4.0
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
Pillow>=10.0.0
numpy>=1.24.0
